[check git] use shell built-in "type" to search for "git"
Use shell built-in "type -a", instead of "which", "where" or "whereis"
as well to check if "git" binary is in $PATH and installed.

In this case, "type" is used with the "-a" option, which returns all of
the places containing the executable file. Furthermore it will also get
functions and aliases.